+++ src/components/TableValues/SkillsPageValues/UpdateSkillForm/UpdateSkillForm.tsx
@@ -10,7 +10,7 @@ import {
   StyledDialogTitle
 } from '../../CvsPageValues/CreateCvForm/CreateCvForm.styles';
 import { IFormInput } from '../../CvsPageValues/CreateCvForm/IFormInput';
-import FormFields from './FormFields';
+import FormFields from '../CreateSkillForm/FormFields';
 
 interface Props {
   opened: boolean;
